minor bugs

+++ src/UptimeCounter.cpp
@@ -50,21 +50,21 @@ unsigned long UptimeCounter::days()
 unsigned long UptimeCounter::hours()
 {
     unsigned long decisecoundsPassed = Rollovers * DS_IN_ROLLOVER + micros() / 100;
-    int hoursPassed = decisecoundsPassed / DS_IN_HOUR;
+    unsigned long hoursPassed = decisecoundsPassed / DS_IN_HOUR;
     return hoursPassed;
 }
 
 unsigned long UptimeCounter::minutes()
 {
     unsigned long decisecoundsPassed = Rollovers * DS_IN_ROLLOVER + micros() / 100;
-    int minutesPassed = decisecoundsPassed / DS_IN_MINUTE;
+    unsigned long minutesPassed = decisecoundsPassed / DS_IN_MINUTE;
     return minutesPassed;
 }
 
 unsigned long UptimeCounter::secounds()
 {
     unsigned long decisecoundsPassed = Rollovers * DS_IN_ROLLOVER + micros() / 100;
-    int secoundsPassed = decisecoundsPassed / DS_IN_SECOUND;
+    unsigned long secoundsPassed = decisecoundsPassed / DS_IN_SECOUND;
     return secoundsPassed;
 }
